refactor merger
